fix: widen Item.value_gp and loot path from u32 to u64 (oag-zrhv3)
PlacedTreasureInstance.gp_value is u64 but action_loot was casting it to
u32 with `as u32`, silently truncating values above ~4.3 billion gp.

- Item.value_gp: u32 → u64
- LootResult.value_gp: u32 → u64
- action_loot explicit_gp param: Option<u32> → Option<u64>
- GM API Loot command value_gp: Option<u32> → Option<u64>
- CLI loot parser: u32 → u64
- Removed truncating `as u32` cast in action_loot

+++ src/command/inventory_cmds.rs
@@ -63,7 +63,7 @@ impl Command for LootCommand {
         let char_name = args[0];
 
         let (item_parts, explicit_gp) = if args.len() >= 3 {
-            if let Ok(v) = args[args.len() - 1].parse::<u32>() {
+            if let Ok(v) = args[args.len() - 1].parse::<u64>() {
                 (&args[1..args.len() - 1], Some(v))
             } else {
                 (&args[1..], None)
